requirements.txt
dlib==19.24.4
opencv-python==4.9.0.80
pillow==10.3.0
face_recognition==1.3.0

